+++ src/components/Detail/DetailComment.jsx
@@ -1,10 +1,11 @@
-import styles from '../styles/ItemDetailComment.module.css';
+import styles from './styles/DetailComment.module.css';
 import { useProductComment } from '@/hooks/useProductComment';
 import { relativeTime } from '@/utils/relativeTimeUtils';
-import ItemDetailProfile from './ItemDetailProfile';
+import DetailProfile from '@/components/Detail/DetailProfile';
+import CommentItemMenu from '@/components/Detail/CommentItemMenu';
 import { useState } from 'react';
 
-export default function ItemDetailComment({ productId }) {
+export default function DetailComment({ productId }) {
   const { productComments, loading, error } = useProductComment(productId, 3, null);
   const [openMenuId, setOpenMenuId] = useState(null);
   
@@ -17,18 +18,14 @@ export default function ItemDetailComment({ productId }) {
   };
 
   return (
-    <ul className={styles.itemDetailComment}>
+    <ul className={styles.detailComment}>
       {productComments.list.map((comment) => (
         <li key={comment.id} className={styles.commentItem}>
           <div className={styles.commentItemDesc}>
             <p className={styles.commentContent}>{comment.content}</p>
-            <CommentItemMenu 
-              isOpen={openMenuId === comment.id}
-              onClick={() => onClickMenu(comment.id)}
-            />
+            <CommentItemMenu isOpen={openMenuId === comment.id} onClick={() => onClickMenu(comment.id)} />
           </div>
-          {/* <p className={styles.commentDate}>{new Date(comment.createdAt).toLocaleDateString()}</p> */}
-          <ItemDetailProfile
+          <DetailProfile
             profileImage={comment.writer.image}
             profileNickname={comment.writer.nickname}
             profileUpdate={relativeTime(comment.updatedAt)}
@@ -39,23 +36,3 @@ export default function ItemDetailComment({ productId }) {
     </ul>
   );
 }
-
-const CommentItemMenu = ({ isOpen, onClick }) => {
-  return (
-    <div className={styles.commentItemMenu}>
-      <button type='button' onClick={onClick}>
-        <img src='/images/common/ic_kebab.svg' alt='메뉴 열기' />
-      </button>
-      {isOpen && (
-        <ul>
-          <li>
-            <button type='button'>수정하기</button>
-          </li>
-          <li>
-            <button type='button'>삭제하기</button>
-          </li>
-        </ul>
-      )}
-    </div>
-  );
-};
