+++ src/components/Detail/styles/DetailComment.module.css
@@ -1,0 +1,26 @@
+.detailComment {
+  display: flex;
+  flex-direction: column;
+  gap: 24px;
+}
+.commentItem {
+  padding-bottom: 12px;
+  border-bottom: 1px solid var(--gray200);
+}
+.commentItemDesc {
+  display: flex;
+  justify-content: space-between;
+  margin-bottom: 24px;
+  p {
+    flex: 1;
+    word-break: break-all;
+    color: var(--gray800);
+    font-size: 14px;
+    font-weight: 400;
+    line-height: 1.71;
+  }
+  button {
+    flex: none;
+  }
+}
+
