+++ src/components/Detail/styles/DetailInquiry.module.css
@@ -1,4 +1,4 @@
-.itemDetailInquiry {
+.detailInquiry {
   margin-bottom: 24px;
   label {
     font-size: 16px;
