+++ src/components/Detail/styles/DetailProfile.module.css
